Type-specifying extension interfaces moved to PHPStan\Type
