+++ src/Type/StaticMethodTypeSpecifyingExtension.php
@@ -1,8 +1,11 @@
 <?php declare(strict_types = 1);
 
-namespace PHPStan\Analyser;
+namespace PHPStan\Type;
 
 use PhpParser\Node\Expr\StaticCall;
+use PHPStan\Analyser\Scope;
+use PHPStan\Analyser\SpecifiedTypes;
+use PHPStan\Analyser\TypeSpecifierContext;
 use PHPStan\Reflection\MethodReflection;
 
 interface StaticMethodTypeSpecifyingExtension
